// frontend/src/app/core/api/model/employeeRequest.ts
/**
 * Le Tacos Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeRequestDto { 
    firstName: string;
    lastName: string;
    phoneNumber: string;
    email?: string;
    salaryAmount: number;
    role?: string;
    employmentStatus: EmployeeRequestDto.EmploymentStatusEnum;
    firstWorkingDay: string;
    emergencyContactName?: string;
    emergencyContactPhone?: string;
    notes?: string;
}
export namespace EmployeeRequestDto {
    export const EmploymentStatusEnum = {
        Active: 'ACTIVE',
        Inactive: 'INACTIVE'
    } as const;
    export type EmploymentStatusEnum = typeof EmploymentStatusEnum[keyof typeof EmploymentStatusEnum];
}